Pass true for the cache_is_immortal flag to the base constructor.
* src/TagStore.cs: Pass true for the cache_is_immortal flag to the
base constructor.

* src/ImportStore.cs (ImportStore.ctor): Pass false for the
cache_is_immortal flag to the base constructor.
* src/PhotoStore.cs (PhotoStore.ctor): Likewise.

* src/Db.cs (DbStore.cache_is_immortal): New private member.
(DbStore.ctor): Get a cache_is_immortal arg.
(DbStore.AddToCache): Add the item if the cache is immortal,
instead of adding a WeakReference to it.
(DbStore.LookupInCache): Expect a WeakReference if the cache is
not immortal, the actual item otherwise.