make auto-creating tag icons optional
When a tag is newly created, it defaults to getting an icon
automatically assigned on first use. Make this optional, with
a checkbox whether an icon should be auto-created, and make
the checkbox remember its state.

Additionally, when an icon was manually removed from a tag,
do not auto-create one again.