Revive the unit tests.
Rework the test build system. This will always build tests if you have nunit
and will ignore it if you don't. Inspiration taken from Banshee's build
system.

The tests are still very verbose on errors and other nasty output (should be
fixed), but most of them are probably broken anyway. However, reviving this
allows for useful tests to be added in the future :-).

Fails distcheck, but I am not capable of fixing it.