Move dbus-sharp, dbus-sharp-glib and dpap-sharp to lib/
All libraries are now in lib, making the build tree a bit nicer. If you
already have a build, you might need to do the following to see the
effects:

	git clean -f -d . -x

This will delete all untracked and ignored files (use with caution),
reverting the tree back to a clean checkout state.