Revert commits 74d4e4..83290d
The new behavior is broken, was reported as such, and still waits for a fix.
In the meantime, I'm reverting as it's annoying.

Revert "Keep the current offsets when a higher quality image gets loaded."
This reverts commit 83290d9488b1abdd33e03cd0e0e6c8335f870eef.

Revert "Preserve the visual size of things when changing image."
This reverts commit 1c347ea470123bcae3c25ed1c9cc019e36608a08.

Revert "No need to compute the minimum zoom every time."
This reverts commit 8b1bcf90b3d331b03624eb6859ef23c214aeb4af.

Revert "Change upscale to a property such that it shouldn't be passed to ZoomFit."
This reverts commit 808007a2624613d38e3966eb23efb177babf3694.

Revert "Change ComputeMinZoom to void."
This reverts commit f71bae535905d294b8138f5d7f0c5d099ea433a0.

Revert "Stop zooming to fit whenever a new pixbuf is pushed in."
This reverts commit 98d84032913a0800e9d0d31712b94499c168167e.

Revert "Refactor image changing API in ImageView."
This reverts commit 74d4e4ae6168ad509807220583d865aabda07486.