Bug 601565 -  Histogram colours don't update on gtk theme change
If the gtk theme of f-spot is changed in the preferences dialog, the colours of
the histogram are not updated. A restart of f-spot is required for this.

This patch at the same time removes the dependency of MainWindow.Toplevel to
make the transition to App.Instance.Organizer possible/easier.