Fix disposed pixmap bug in Adjust Colours tool
The code paths in Adjustment.Adjust () for the alpha- and non- alpha
channel cases have an important difference - the alpha channel codepath
modifies Input and returns it, the non-alpha channel path returns a
modified copy of Input.

In the alpha case, returning Input eventually results in the preview
pixmap being disposed, causing debug spew and (in View mode) the window
to display garbage.