Clean-up build
This commit switches most of the build to use the banshee stuff in build/.
Only the libraries are just left out for now.
The file src/PixbufUtils.cs is changed, because the build environment
checks for consistency of DllImports. This check failed for this file
because in input was there in a commented part of the code.

https://bugzilla.gnome.org/show_bug.cgi?id=620379