Big leap closer to Taglib# usage.
WARNING: this commit causes file corruption! See below!

Changes:
* Migrated from PixbufOrientation to TagLib.Image.ImageOrientation.
* Parse dates and descriptions directly using TagLib, instead of in
  ImageFile.
* Move metadata writing mostly to TagLib.
* Added (non-working) unit test for GIOTagLibFileAbstraction.

There seems to be an interesting corruption happening when using the
GIOTagLibFileAbstraction. This does not occur when using normal IO. Am
puzzled, but I suspect it is something related to the read/write
behavior of GIO that trips this up.

Apparently GIO has atomic write support, which we should use. Needs
fixing.