Split up IconView in smaller classes
IconView is split up in the classes CollectionCellGridView and SelectionCollectionGridView.
CollectionCellGridView adds the drawing of thumbnails to the basic CellGridView. Some
render code is moved out to the CaptionRenderer to make it more configureable.
SelectionCollectionGridView adds a SelectionCollection which allows to select an arbitrary
number of photo. The code for the rubberband selection is mostly copied, but it needs to
be cleaned up also.