Fixes bgo#670498 f-spot crashed - Export -> CD
Not exactly sure why it was crash, but a simple
change to how the foreach was handled seems to
fixed it.

Also some cleanups wrt FSpot.Exporter.CD.  I
renamed some methods to be easier to understand
and improved some of the code.  The hope is with
the addition of IDiscBurner that eventually other
platforms/apps can be used.