UI integration and issue fixes.
- FacesToolWindow is no longer a floating window, now it is in the
sidebar —to integrate better with F-Spot UI.

- To achieve the above I had to improve the way of instantiating
FacesTool: Before this change you needed to press a button to
activate the tool, but now it will be always visible when displaying
a single image, so I had to connect to some events to instantiate
it in the correct time to avoid some issues like copying a buffer
of low quality that is shown while the image is loading or a
problem with faces locations because of instantiating FacesTool
before the size allocation for the image.

- I've extended the changes needed for the above to correct a big design
problem: The face shapes need an scaled version of the pixbuf
currently displayed to be able to *erase* themselves, but until now
each of the shapes maintained a scaled copy, wich was of course more
time and memory consuming than needed. Now that scaled copy is
maintained by FacesTool.