UI integration improvements.
- There was a lot of problems when switching from having a floating window
to showing the tool in the sidebar, like loading incorrect face locations
when moving to next/previous image and having the FacesTool always
active so you couldn't use other tools that require selection. Now
the FacesTool has a better activation and loading mechanism to avoid
unnecesary load of data when the tool is not active. It will be
active if you have *Faces* selected in the sidebar page selector,
and if you don't, the face locations won't be shown when entering
the image area with the mouse, and of course the selection works
again. Also, if you change to next/previous image the FacesTool
won't be loading data if it isn't visible.

- The *Cancel* button has been removed because now there isn't a dialog
to close, and the *OK* button has been replaced by a *Save* button
that will save the changes if there is something to save.