Fix Hyena exception on first start of f-spot when no photos.db exists.
For newly created databases the code surrounded by the try/catch will
always cause an exception as no tags table exists at this point in
time. The hidden tag is created later in the startup sequence by
TagStore class.

The meta table was introduced between versions 0.1.8 and 0.2.0. So for
*very* old databases the meta table needs to be created from scratch.
In this case the code is required to also create the entry for the
hidden tag.

We distinguish between these cases by checking for the existence of
the tags table and therefore remove the exception.