[Build] Add ICSharpCode.SharpZipLib as a nuget.
More stuff is moving to nuget. It unfortunate that it isn't very distro
packaging friendly. Mono-addins moved to nugets as well and since we
import that and our tests already use nugets I figured we'd just finally
jump in with both feet.

This also makes it slightly easier to develop and who knows... one day
run on Windows (and Mac?)?